chore: 메인페이지 prompt 페이지로 변경
- 책 보관함 페이지는 bookshelf페이지로 디렉토리 명 변경
